src/requirements.txt
Flask==2.3.3
PyMySQL==1.1.0
mysql-connector-python==8.1.0
gunicorn==21.2.0